# src/routes/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0",
      "title": "sobermart",
      "contact": {}
    },
    "host": "kuro.asrofur.me",
    "basePath": "/sober/api",
    "securityDefinitions": {},
    "schemes": [
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/{main}{auth}/check": {
        "get": {
          "summary": "CheckToken",
          "tags": [
            "Auth"
          ],
          "operationId": "CheckToken",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "main",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "auth",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "security": []
        }
      },
...
    },
    "security": [],
...
  }